Update: Moved components to middleware file
